% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{calcBlockCoverage}
\alias{calcBlockCoverage}
\title{calcBlockCoverage}
\usage{
calcBlockCoverage(object, min.block = 5, ncores = 16)
}
\arguments{
\item{object}{an object of meth class}

\item{min.block}{minimum number of CpG sites per block}

\item{ncores}{number of cores to use}
}
\value{
an object of meth class
}
\description{
calculates block statistics for analysis
calculates average number of CpGs per block (block_mean)
calculates total number reads per block (block block_sum)
calculates number of CpGs per block (L)
calculates average methylation level per block (PC)
calculates average impution per block (imp)
}
