% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{createIndex}
\alias{createIndex}
\title{createIndex}
\usage{
createIndex(object)
}
\arguments{
\item{object}{a meth class object}

\item{input_dir}{the directory of the count and methylation count files}

\item{achr}{is the name of the chromosome for example chr17}
}
\value{
A meth class object
}
\description{
This function creates converts rownames chr17.1002 to a Granges object
Ensures that the rownames are equal and sorted
Generates rowsums for each CpG site
Generated rowmeans for each CpG site
Results are stored in object
The naming convention for the coverage file is  chr??_tot_cov.tab
The naming convention for the methylation file is chr??_tot_meth.tab
Files are table delimited
The rownames must be in the following format chr.pos for example chr17.10002
The colnames contains are the sample identifiers.
Duplicate rownames are not allowed
Modified total coverage is stored in object@raw_reads
Modified Total methylated is stored in object@raw_meth
Granges index is stored in object@index
}
