% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockClassDefinition.R
\name{createMethObject}
\alias{createMethObject}
\title{createMethObject}
\usage{
createMethObject()
}
\value{
an meth class object
}
\description{
This constructs a meth class object with attributes
}
