% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{createSegIndex}
\alias{createSegIndex}
\title{getSegIndex}
\usage{
createSegIndex(segs, achr)
}
\arguments{
\item{segs}{a list of segments}

\item{achr}{a character chromsome name ie chr17}
}
\value{
list with segment names as keys and GRange as values
}
\description{
Helper function creates list of segment names and corresponding GRange
}
