% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{endreCluster}
\alias{endreCluster}
\title{endreCluster}
\usage{
endreCluster(bb = bb, bwd = 300, min.block = 3, hclust = 0.4)
}
\arguments{
\item{bb}{a segment}

\item{bwd}{bandwidth}

\item{min.block}{minumum number CpGs sites per block}

\item{hclust}{hclust cutoff for dividing blocks}
}
\value{
a list of blocks per segment
}
\description{
Divides segments into blocks using methyDist and hclust
Returns list of blocks per segment
Filters out blocks with less than min.block CpG sites
}
