% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{filterIQR}
\alias{filterIQR}
\title{filterIQR}
\usage{
filterIQR(object, iqr_cutoff = 10)
}
\arguments{
\item{object}{an object of meth class}

\item{iqr_cutoff}{the iqr cutoff}
}
\value{
an object of meth class
}
\description{
Filters blocks whose maximum-minimum divided by iqr is greater than cutoff
This removes blocks were one or two sites are responsible for variation
Updated index stored in object@index
Updated blocks stored in object@blocks
}
