% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{getBlocks}
\alias{getBlocks}
\title{getBlocks}
\usage{
getBlocks(object, bwd = 300, min.block = 5, hclust = 0.4, ncores = 16)
}
\arguments{
\item{object}{an object of meth class}

\item{bwd}{bandwidth}

\item{min.block}{minumum number CpGs sites per block}

\item{hclust}{hclust cutoff for dividing blocks}

\item{ncores}{number of cores to use}
}
\value{
a list of blocks for all segments on a chromsome
}
\description{
Iterates over all segments
Divides all chromsome segments into blocks based on covariation and distance
Limits index to blocks
Blocks are stored in object@blocks
Updated index stored in object@index
}
