% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{getGRangeIndex}
\alias{getGRangeIndex}
\title{getGrangeIndex}
\usage{
getGRangeIndex(xx, achr)
}
\arguments{
\item{xx}{a vector of characters ie c("chr17.10001, chr17.10002, ...)}

\item{achr}{a chromosome name "chr17"}
}
\value{
a sorted Rrange
}
\description{
Helper function for converting row.names ie chr17.10001 to Granges
}
