% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{getReturnData}
\alias{getReturnData}
\title{getReturnData}
\usage{
getReturnData(object)
}
\arguments{
\item{object}{an object of meth class}
}
\value{
list of m2=imputed matrix, blocks=blocks, index=index, raw_cov=raw_cov and segments=segs
}
\description{
Returns list of m2=imputed matrix, blocks=blocks, index=index, raw_cov=raw_cov and segments=segs
}
