% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{getSegRanges}
\alias{getSegRanges}
\title{getSegRanges}
\usage{
getSegRanges(segs)
}
\arguments{
\item{segs}{}
}
\value{
a table of segment with methylation events and length
}
\description{
Helper function for converting segments to Granges
Adds number of methylation events and length
expects a list of segments
}
