% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{getSegs}
\alias{getSegs}
\title{getSegs}
\usage{
getSegs(object, min.seg = 5)
}
\arguments{
\item{object}{of class meth}

\item{min.seg}{is the minimum number of CpG sites needed to keep the segment}
}
\value{
an object of meth class
}
\description{
This function divides the chromome into segments based on distance
This function filters out segment less than min.seg parameter
Distance paramater is the closest distance to 3000 bp
Segments are stored in the as GRanges in object@segs
}
