% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{methylDist}
\alias{methylDist}
\title{methylDist}
\usage{
methylDist(X, pos, bwd = 300)
}
\arguments{
\item{X}{a segment}

\item{pos}{numeric postion of methylation sites on segment}

\item{bwd}{bandwidth}
}
\value{
a matrix generated from covariation, and weighted by distance
}
\description{
Calculates a correlation matrix of segment relative methylation d1=cor(seg)
Calculates a distance matrix by segment position and weighted by bandwidth
d2 = exp(-as.matrix(dist(pos)^2)/(2\emph{bwd^2))
Returns d, where d=1 - d1}d2
}
