% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_methods.R
\name{ReadRawFiles}
\alias{ReadRawFiles}
\title{Load coverage and methylation files}
\usage{
ReadRawFiles(object, input_dir = "./", achr = "chr17")
}
\arguments{
\item{object}{a meth class object}

\item{input_dir}{the directory of the count and methylation files}

\item{achr}{is the name of the chromosome for example chr17}
}
\value{
A meth class object
}
\description{
This function loads two files (total coverage and total methylated).
The coverage tot_cov.tab is the number of reads at each CpG site
The methylation file is the number of methylated reads at each CpG site
The naming convention for the coverage file is  chr??_tot_cov.tab
The naming convention for the methylation file is chr??_tot_meth.tab
Where ?? is the chrosome (chr17) and input directory(./data)
Files must be tab delimited
The rownames must be in the following format chr.pos for example chr17.10002
The colnames contains are the sample identifiers.
Duplicate rownames or colnames are not allowed
Chromsome is stored in object@chr
Total coverage is stored in object@raw_cov
Total methylated is stored in object@raw_meth
}
