% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlocks.R
\name{runChromosome}
\alias{runChromosome}
\title{runChromosome}
\usage{
runChromosome(
  achr = "chr18",
  input_dir = "./inst/extdata/",
  min.seg = 20,
  bwd = 300,
  min.block = 5,
  hclust = 0.3,
  iqr_cutoff = 10,
  nb = 5,
  ncores = 16
)
}
\arguments{
\item{achr}{a name of a chromosome for example chr18}

\item{input_dir}{directory containing chr18_tot_cov.tab and chr18_tot_met.tab files}

\item{min.seg}{is the minimum number of CpG sites needed to keep the segment}

\item{bwd}{or distance decay bandwith parameter}

\item{min.block}{minimum number of CpG site per block}

\item{hclust}{the cutoff used in hclust to divide into blocks}

\item{iqr_cutoff}{the range/iqr to remove blocks with 1-2 outliers}

\item{nb}{number of neighbours used by KNN to impute missing values}

\item{ncores}{number of cores to use}
}
\value{
a list of results including blocks with statistics
}
\description{
Main function for finding blocks on a chromosome
}
